[add] add a type size_t

+++ src/declare.c
@@ -19,7 +19,7 @@ Type *set_type(Type *type, Token *tok){
 		case BOOL:
 		case CHAR:
 		case INT:
-		case SIZE_T:
+		case LONG:
 		case PTR:
 		case ARRAY:
 			break;
@@ -108,8 +108,8 @@ Type *parse_type(void){
 	}else if(consume_reserved_word("int", TK_TYPE)){
 		type->ty = INT;
 		type = set_type(type, NULL);
-	}else if(consume_reserved_word("size_t", TK_TYPE)){
-		type->ty = SIZE_T;
+	}else if(consume_reserved_word("long", TK_TYPE)){
+		type->ty = LONG;
 		type = set_type(type, NULL);
 	}else if(consume_reserved_word("struct", TK_TYPE)){
 		type->ty = STRUCT;
